Get angles between average cluster direction and cluster-supporting lines

Determine the angles between the average cluster direction and the cluster-supporting lines. These angles are obtained from a wrapped normal distribution (\(\mu=0\), \(\sigma=\) angle_disp ) with support in the interval \(\left[-\pi/2,\pi/2\right]\). Note this is different from the standard wrapped normal distribution, the support of which is given by the interval \(\left[-\pi,\pi\right]\).

Usage

angle_deltas(num_clusters, angle_disp)

Arguments

num_clusters

Number of clusters.

angle_disp

Angle dispersion, in radians.

Value

Angles between the average cluster direction and the cluster-supporting lines, given in radians in the interval \(\left[-\pi/2,\pi/2\right]\)

Note

This function is stochastic. For reproducibility set a PRNG seed with set.seed.

Examples

set.seed(123)
arad <- angle_deltas(4, pi / 8) # Angle dispersion of 22.5 degrees
arad                            # What angles deltas did we get?
#> [1] -0.22009827 -0.09039049  0.61210332  0.02768858
arad * 180 / pi                 #  Show angle deltas in degrees
#> [1] -12.610702  -5.178994  35.070937   1.586439