clugenr: Multidimensional Cluster Generation Using Support Lines

An implementation of the clugen algorithm for generating multidimensional clusters with arbitrary distributions. Each cluster is supported by a line segment, the position, orientation and length of which guide where the respective points are placed. This package is described in Fachada & de Andrade (2023) doi:10.1016/j.knosys.2023.110836 .

See also

Author

Maintainer: Nuno Fachada faken@fakenmc.com (ORCID) [copyright holder]